pyproject.toml
# ==========================================================
# 构建系统配置
# 这部分告诉打包工具（如pip, uv）如何构建这个项目。
# 这是让 entry_points 生效所必需的。
# ==========================================================
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"} # 告诉setuptools，包的根目录在'src'下

# ==========================================================
# 项目元数据 (由 uv 管理)
# ==========================================================
[project]
name = "stock-downloader"
version = "1.6.0"
description = "一个基于Tushare Pro的、基于DuckDB的个人量化数据下载器。"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.3.2",
    "graphviz>=0.21",
    "pandas>=2.3.1",
    "tqdm>=4.67.1",
    "tushare>=1.4.21",
    "typer>=0.12.3",
    "pyrate-limiter>=3.9.0",
    "python-box>=7.3.2",
    "tomli-w>=1.2.0",
    "huey>=2.5.3",
    "dependency-injector>=4.48.1",
    "pyarrow>=21.0.0",
    "poethepoet>=0.37.0",
    "ibis-framework[duckdb,examples]>=10.8.0",
    "toml>=0.10.2",
]

[project.scripts]
neo = "neo.main:app"



# ==========================================================
# Ruff configuration
# ==========================================================
[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = ["E501"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Allow E402 (module level import not at top of file) for notebooks
# since initialization code may need to run before imports
# Also allow for scripts that need to modify sys.path before imports
[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E402"]
"debug_progress.py" = ["E402"]
"examples/*.py" = ["E402"]

# ==========================================================
# Coverage configuration
# ==========================================================
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py", 
    "*/interfaces.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]

# ==========================================================
# MyPy configuration
# ==========================================================
[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "tushare.*",
    "duckdb.*",
    "plotly.*",
    "tabulate.*",
    "pyrate_limiter.*",
]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest-cov>=6.2.1",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.12.7",
    "pytest>=8.4.1",
    "mypy>=1.11.0",
    "types-pyyaml>=6.0.12.20250516",
    "ipykernel>=6.30.1",
    "pipdeptree>=2.28.0",
    "ibis-framework[duckdb,examples]>=10.8.0",
]

# ==========================================================
# Poethepoet 任务管理配置
# ==========================================================
[tool.poe.tasks]

# 测试相关任务
test = "uv run python -m pytest"
test-cov = "uv run python -m pytest --cov=src --cov-report=term-missing"
test-fast = "uv run python -m pytest -x --tb=short"
test-verbose = "uv run python -m pytest -v"

# 代码质量检查
lint = "uv run ruff check src tests"
lint-fix = "uv run ruff check --fix src tests"
format = "uv run ruff format src tests"
format-check = "uv run ruff format --check src tests"
type-check = "uv run mypy src"

# 组合任务
check = ["lint", "format-check", "type-check"]
fix = ["lint-fix", "format"]
ci = ["check", "test-cov"]

# 项目运行任务
run = "uv run neo"
run-help = "uv run neo --help"

# 开发工具
clean = "rm -rf .pytest_cache .mypy_cache .ruff_cache __pycache__ .coverage"
deps = "uv run pipdeptree"

# 脚本任务
redown = "uv run python scripts/check_and_redown.py"
monitor = "uv run python scripts/huey_monitor.py"
qdata = "uv run python scripts/query_parquet_data.py"
msync = "uv run python scripts/manual_sync_metadata.py"

# Huey 相关任务
huey-fast = "uv run neo dp fast"
huey-slow = "uv run neo dp slow"

[tool.poe.tasks.huey-all]
shell = """
trap 'echo "\nShutting down Huey consumers..."; kill 0' EXIT
poe huey-fast &
poe huey-slow
"""
help = "Starts both Huey consumers and the monitor, with graceful shutdown."
